Refactored equals() and hashCode() function check
